+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -5,5 +5,5 @@ import ru.yandex.practicum.filmorate.model.User;
 public interface UserStorage {
     User addUser(User user);
     User updateUser(User user);
-    void removeUser(int id);
+    void removeUser(long id);
 }
